api-gateway/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
requests>=2.31.0
python-multipart>=0.0.6
aiofiles>=23.2.1
Pillow==10.0.1 